feat: use element typname appended with [] to identify arrays
Appending [] to the element type name is consistent with what is
expected as input and with PostgreSQL behavior. It's also more robust
as the database-generated array type names are creation-order
dependent. Names with _ prefix are still accepted as input and match
types in pg_type that are named that way. This is confirmed by tests.

This does cause a few existing tests to break. These will be updated
in a separate commit.