feat: standardize TypeInfoCache typename input and output
This patch standardizes the format of type names parsed and emitted by
TypeInfoCache.

Type name strings parsed and emitted by TypeInfoCache have not been
consistent and have suffered from a number of issues:

 - Some valid PostgreSQL type names could not be specified.
 - Some type names returned by getPGType(oid) could not be used as
   input to getPGType(name). Sometimes the name could not be parsed,
   other times it would match a different type than the one returned.

Type names in general follow the PostgreSQL identifier quoting
conventions. Double quotes are used to enclose identifiers containing
special characters, and quotes within quoted identifiers must be
escaped with a quote. All valid PostgreSQL type names are accepted.

One difference is that quotes are only required under the following
situations:
 - the schema or type name contains a dot
 - the schema or type name starts with a quote
 - the type name ends in "[]"

Type names are supplied to and returned from TypeInfoCache methods
like `createArrayOf` and `getColumnTypeName` as Java String
values. These are in isolation of a SQL statement such as "SELECT
'foo'::text" where PostgreSQL will perform casefolding. Other database
metadata methods in JDBC return case-sensitive identifiers unquoted,
and it makes sense to do so here as well. Likewise, it's desireable to
be able to round-trip values returned by `getColumnTypeName` as
arguments to `setType` and `createArrayOf`. As such, quotes are not
needed (but are accepted) for case sensitivity or other special
characters and identifiers are not casefolded. For example: "My.Enum"
and "\"My.Enum\"" are both valid type name representations of the
PostgreSQL type "My.Enum".

For ease of use, one exception to this is the case of unqualifed
identifiers. Unqualified identifiers will also be lowered and checked
against pg_catalog: "TEXT" and "text" will both match the PostgreSQL
text type.

   input      | schema     | type        | emitted as
  ------------|------------|-------------|------------
   text       | pg_catalog | text        | text
   TEXT       | pg_catalog | text        | text
   foo        |            | foo         | foo
   FOO        |            | FOO         | FOO
   "foo"      |            | foo         | foo
   foo[]      |            | foo array   | foo[]
   "foo[]"    |            | foo[]       | "foo[]"
   "foo[]"[]  |            | foo[] array | "foo[]"[]
   f"oo       |            | f"oo        | f"oo
   "f""oo"    |            | f"oo        | f"oo
   "foo       | parse error
   """foo"    |            | "foo        | """foo"
   f.oo       | f          | oo          | f.oo
   "f.oo"     |            | f.oo        | "f.oo"
   ns.foo     | ns         | foo         | ns.foo
   "ns"."foo" | ns         | foo         | ns.foo
   NS.FOO     | NS         | FOO         | NS.FOO